+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- The Button component now supports is-dense.
+- User supplied Button classes are now inserted after those added by the component.
+
 ### Changed
 
 ### Removed